dev_ops/hw1/preprocess_service/requirements.txt
fastapi
uvicorn
pillow
numpy
torch
python-multipart
prometheus-fastapi-instrumentator